+++ src/SophiaWin11.Core/Abstractions/IAnimationService.cs
@@ -1,8 +1,0 @@
-namespace SophiaWin11.Core.Abstractions;
-
-public interface IAnimationService
-{
-    bool AnimationsEnabled { get; }
-
-    void PlayIntro();
-}