+++ src/SophiaWin11.Core/Abstractions/IThemeService.cs
@@ -1,8 +1,0 @@
-namespace SophiaWin11.Core.Abstractions;
-
-public interface IThemeService
-{
-    bool IsThemeApplied { get; }
-
-    void ApplyTheme();
-}